refactor(playbook): extract engine.py into modular package
Split the monolithic engine.py (830 lines) into focused modules:
- context.py   : NamedList, MissingDataTracker, _flatten, _build_context
- conditions.py: evaluate_condition() helper + _stringify_condition()
- templates.py : Jinja2 env, filters, _resolve_template(), _resolve_path()
- loader.py    : load_playbook()
- sections.py  : _evaluate_section() and sub-functions
- integrations/gitlab.py: labels, checklists, templates resolvers
- evaluator.py : clean evaluate() orchestrator (~160 lines)
- engine.py    : backward-compatible re-export proxy (58 lines)

All existing imports from regis_cli.playbook.engine are preserved.
218 tests pass, ruff clean.
