+++ regis_cli/playbook/integrations/__init__.py
@@ -1,0 +1,1 @@
+"""GitLab integration package for regis-cli playbook engine."""
